<AVR> Added arch suspend/resume signal api
These apis are to be used when entering and exiting suspend state.
While entering cpu suspend state, programmer is intended to call
and store the state, so that upon resume, certain checks for boot
api needs be performed based on which they need to be executed.
i.e.; when cpu enters simple wfi sleep, upon resume cpu need not
perform full platform boot, (assuming cpu is reset; not applicable
for AVR core, but these are generic apis)

Issue: #169