<RISC-V> Added stack pointer update api
This function to be used only where there is not critical context.
Upon calling this function the old context will be permanently lost
and stack will be initialised to newer location without ensuring
context management.
Note: It is recommended to use this function only in naked function,
the one which don't person cxt-mgnt.

Issue: #169